+++ packages/components/CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `MenuList` unnecessary windowing
+- `MenuGroup` sticky group label stacking on top of items
 - `Tooltip` shows on an input with `autoFocus` on first mount
 - `FieldInline` positioning correction correction. Affects:
   - `FieldCheckbox`